requirements.txt
# Giggso Build-Break Challenge — Incident Knowledge Assistant
# Python 3.10+

# Frontend & server
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Bot / LLM integration (open-source fallback)
requests>=2.31.0
httpx>=0.25.0

# Voice: speech-to-text (open source)
SpeechRecognition>=3.10.0
PyAudio>=0.2.14
# Optional: gTTS for text-to-speech (online), pyttsx3 for offline TTS
gTTS>=2.4.0
pyttsx3>=2.90

# Environment and auth
python-dotenv>=1.0.0

# Utilities
python-multipart>=0.0.6
